[update]install - Fix incorrect installation of git bashrc.

+++ install.sh
@@ -11,27 +11,28 @@ do
     ln -s ${PWD}/${tool}.sh ${PWD}/${BIN_DIR}/${tool};
 done
 
+export PATH=${PWD}/bin:${PATH}
+
 # Install configurations of vim
 echo "Installing VIM-related configurations";
 tar xjf vim-related.tar.bz2
-mv ~/.vim ~/.vim_bak
-mv ~/.vimrc ~/.vimrc_bak
-mv .vim* ~
+if [[ -e ~/.vim ]];then
+  mv ~/.vim ~/.vim_bak
+fi
+if [[ -e ~/.vimrc ]];then
+  mv ~/.vimrc ~/.vimrc_bak
+fi
+mv my.vim ~/.vim
+mv my.vimrc ~/.vimrc
 
 # Install git-supported .bashrc
 echo "Installing git-related .bashrc";
 tar xjf git_bashrc.tar.bz2
 if [[ -e ~/.bashrc ]];then
-  md5_current=$(md5sum ~/.bashrc | awk '{print $1}');
-  md5_new=$(md5sum .bashrc | awk '{print $1}');
-  if [ $md5_current = $md5_new ];then
-    echo INFO: same .bashrc;
-  else
-    mv ~/.bashrc ~/.bashrc_bak;
-    mv .bashrc ~
-  fi
-else
-  mv .bashrc ~
+  mv ~/.bashrc ~/.bashrc_bak;
 fi
+#mv .bashrc ~
+echo "export PATH=${PWD}/bin:${PATH}" > ~/.bashrc
+cat my.bashrc >> ~/.bashrc
+rm my.bashrc
 
-export PATH=${PWD}/bin:${PATH}
